pyproject.toml
# AI聊天系统项目配置
[project]
name = "ai-chat-system"
version = "1.0.0"
description = "统一AI聊天系统 - 支持多AI服务提供商的聊天接口"
readme = "README.md"
requires-python = ">=3.13"

# 项目依赖
dependencies = [
    # FastAPI 核心
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # 数据库
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",  # PostgreSQL 异步驱动
    "aiomysql>=0.2.0",  # MySQL 异步驱动
    "alembic>=1.13.0",  # 数据库迁移

    # 缓存和消息队列
    "redis>=5.0.0",
    "aioredis>=2.0.0",

    # HTTP 客户端
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",

    # 认证和安全
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",

    # 工具库
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",  # 重试机制
    "orjson>=3.9.0",    # 高性能JSON
]

# 开发依赖（可选）
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
